Remove @CompileTimeConstant (#74)
In practice, it can be tricky to make sure that everything is a CompileTimeConstant. Rather than demanding this at the lowest levels of our stack, we should be enforcing appropriate behaviour higher up. In particular, we would like to provide this functionality by default in an internal library, but the compile-time nature of this cannot be guaranteed by error-prone through a chain of lambdas.