[fix] Don't use parent span operation in deferred traces (#110)
## Before this PR
When creating a deferred trace without specifying an operation, the parent span operation is used. This is confusing because:
1. It creates duplicate spans with the same operation name
2. The parent span operation name is likely not a good description of what the deferred trace task is doing.

## After this PR
When creating a deferred trace without specifying an operation, the default deferred span operation is used. The constructors that allow callers to not specify an operation have been deprecated a signal to clients that they should provide an operation for better trace logs.