[improvement] Set the default sampler to random with 1% (#158)
## Before this PR
The default sampler would sample all traces. I would claim this is not a great default and is somewhat error prone. It expects all consumers to override it to an appropriate value, with the penalty to failing to doing so being a high number of tracing logs that puts a lot of pressure in logging pipelines.

## After this PR
The default sampler is random with 1% probability of sampling.