[improvement] Allow setting observable flag when wrapping with new trace (#163)
<!-- PR title should start with '[fix]', '[improvement]' or '[break]' if this PR would cause a patch, minor or major SemVer bump. Omit the prefix if this PR doesn't warrant a standalone release. -->

## Before this PR
<!-- Describe the problem you encountered with the current state of the world (or link to an issue) and why it's important to fix now. -->
`Tracer.wrapWithNewTrace` and `Tracer.wrapWithAlternateTraceId` always initializes the new trace with `Observability.UNDECIDED`, and there is no easy way to force a particular trace to be sampled.

## After this PR
<!-- Describe at a high-level why this approach is better. -->
Added overloaded versions of `wrapWithNewTrace` and `wrapWithAlternateTraceId` that take an additional parameter to allow users to set whether the new trace should be observable.
<!-- Reference any existing GitHub issues, e.g. 'fixes #000' or 'relevant to #000' -->

This is useful to us because we have a service that makes requests to two separate backends, one of which is wrapped using `wrapWithNewTrace` so that it generates a separate trace, and we want the secondary trace to be sampled when the primary trace is sampled in order to compare the traces. This should become possible with this PR and https://github.com/palantir/tracing-java/pull/164.