Request log tracing
At the moment, we send trace ids and span ids with our requests. This is
excellent, and provides some coarse grained information. That said, it's
very limited - you can merely know that some number of requests were
triggered by the same underlying operation.

Here, when a trace is originated with a parent span, we keep track of
that parent (called an 'originating' span id) and also send it with
requests. This means that we can use request logs to piece together
accurate service level tracing as well as trace logs.

In other words, if I have a call graph which looks like:

```yml
service a/foo:
   check auth:
     call auth service:
       auth service/check auth:
   do work:
     call work service:
       work service/do operation:
         call auth service:
           auth service/check auth:
```

then I will see that in the trace logs (sampled) but I can derive

```yml
service a/foo:
   auth service/check auth:
   work service/do operation:
     auth service/check auth:
```

from the request logs on all requests. Currently this is possible
when requests are single-threaded, but not when people call services in
parallel.