Introduce localTraceId to disambiguate incoming requests that share traceid
Logs from concurrent incoming requests with the same traceid are currently difficult
to tell apart (one can look at thread names etc, but not across executors).

This introduces a new optional property on a Trace, localTraceId, that we set to
a new unique identifier if an incoming request already has a traceid assigned.