Provide Tracing support for the full request lifecycle
Implement TracedRequestHandler which can be applied prior to blocking
operations in order to capture more of the request lifecycle.
The existing handler attaches the detached span created by the
request handler in order to provide more specific timing information
describing the blocking portion of the request.

For backwards compatibility, the existing TracedOperationHandler
can still be used without the TracedRequestHandler, however the
detached request span will be started when the TracedOperationHandler
is first executed.