Tracing no longer depends on afterburner
Afterburner was used by the AsyncSlf4jSpanObserver, however that
observer implementaiton is never used in production and has
serious performance issues at scale. In the future, we should split
the observer implementation into a separate module to decouple the
tracing library from jackson modules (with the excpetion of
annotations, perhaps).