Allow conjure-undertow to leverage the `DetachedSpan.attach()` mechanism
This allows us to remove a largely redundant tracing span and flatten
our graphs very slightly.

Note that the sampled span warn-if-completed logging has been
removed. I think adding it originally was a mistake given
the probabilistic nature. There are many scenarios in which
a server exchange may end prior to code execution (for example
when clients disconnect) where we would expect to see these
warnings.

This leads us to a potential downside: When clients disconnect
before the server finishes processing blocking work, the spans
may provide somewhat less information. This is an edge case,
and I don't believe it's worth the additional spans, but is worth
describing here.