Add Jakarta EE9 Equivalents of all libraries (#962)
Add Jakarta EE9 Equivalents of all libraries

This is the second attempt at this. This version however is
significantly simpler than the previous version since we had
already converted all of the tests to a simpler undertow only
implementation, removing the Dropwizard dependency which
inadvertently bumped all of our dependencies past compatible
versions.

Tests have been consciously dropped in the non-jakarta versions as
the dependencies get messed up with those.

Similarly, revapi was dropped from the non-jakarta versions because of
a quirk in how that plugin works. It's not providing value here anyways though.