Tracing-aware executors invoke the uncaught exception handler while a trace is set
Previously the tracer state would be cleaned up prior to the exception escaping
scope, so it would not have a relevant traceId/etc configured.

The trade-off is that this could introduce risk when executors are
composed, listening for uncaught exceptions. nylon-threads, for example,
must be updated as it expects to capture such throwables, but doesn't
explicitly update thread state:
https://github.com/palantir/nylon-threads/blob/aa361f56b91bebcdc477a198c2e8fa6fd36f554f/nylon-threads/src/main/java/com/palantir/nylon/threads/RenamingExecutorService.java#L91-L94