Overhaul span stack
I've realized that we probably need to allocate and push an entry for disabled spans.

If you create a disabled span and then start it on another thread, we still want the trace information to propagate. This is a generalization of the problem that motivated the introduction of defaultTraceState.

It's hard to avoid work here because at span creation time we don't know whether a span is going to be moved to a different thread before being attached.

Other tracing implementations avoid this problem but make other tradeoffs:
- tracing_opentelemetry is limited by the tracing APIs, which does not have trace IDs. So changing the span filter affects the trace IDs for spans (ie. disabling a span sent to another thread prevents the trace ID from propagating to that other thread.
- opentelemtry_sdk doesn't support moving spans before they are attached (spans are started and attached immediately upon creation).